openchangedb: Impelement Logger module
Important part is openchangedb.c:openchangedb_initialize()
For now it is very simple and not flexibble, but this approach
will allow us to 'stack' backend chains.

openchangedb_logger is an example for doing staff during backend calls.
My aim is to make it configurable through loadparm_context like:
  mapiproxy:openchangedb = logger,mysql
Other thing would be to have a simple registry for backend modules,
so we could get rid of this clumsy if (myslq)... else if(ldb)...